sdk/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentflow"
version = "0.1.0"
authors = [
  { name = "AgentFlow Team" },
]
description = "Python SDK for AgentFlow - Kubernetes for AI Agents"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.25.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.8.0",
    "pyyaml>=6.0",
    "asyncio-nats-client>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "flake8>=5.0.0",
]

[project.urls]
"Homepage" = "https://github.com/Siddhant-K-code/agentflow-infrastructure"
"Bug Reports" = "https://github.com/Siddhant-K-code/agentflow-infrastructure/issues"
"Source" = "https://github.com/Siddhant-K-code/agentflow-infrastructure"